Move write_then_drop tests into tcp test file
